fix(rh-widgets): Throw an error if RH widget queries are beyond July 12
Error if RH widget queries are beyond Jul 12.
